backend/requirements.txt
# Carotid segmentation (train script + notebook)
torch>=2.0.0
monai>=1.3.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
PyWavelets>=1.4.0
jupyter>=1.0.0
notebook>=7.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
joblib>=1.3.0

# Backend: FastAPI · SQLAlchemy · Pydantic v2 · JWT · Firebase
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
firebase-admin>=6.4.0
python-multipart>=0.0.9
africastalking>=1.2.0
